Make cirrus unit tests run with cirrus
Fix two problems with nose to enable cirrus unit
tests to be run with `git cirrus test`.

The function `nose_test` in the test module is
decorated to tell nose that it is not a test which
otherwise causes an error.

The unit test module names are changed so that they
match the nose regexp for finding test modules.