Rewrite github_tools functions not needing github
Move functions in github_tools that do not relate to github
specific features to git_tools, and rewrite these functions to
provide (nearly) the same functionality using the GitPython package
rather than the github API and requests.

This allows new releases to be created without a github token or
even a github account. Previously, the only functionality in making a
new release requiring the github API was for `git cirrus release new`
to retrieve all commits since the last release and add records of
these commits to `HISTORY.md`.

The old method for building `HISTORY.md` was apparently not working
correctly, as github was returning the last 30 commits, rather than
all commits since the last release. This bug is fixed with the changes.

Additionally, the old function `get_tags` was sorting the tags in
alphabetical order by tag name rather than by newest first. This would
cause version `0.1.13` to appear as older than `0.1.9`. This bug is
also fixed with the new function.

Further checking of this rewriting and unit tests will follow.