Unit tests and github_tools fixes
Rewrite unit tests for functions moved from github_tools
to git_tools. These tests could have better coverage, but the
purpose of this commit is to at least restore any coverage
that was previously present.

Functions in github_tools that call functions previously in
git_tools have been repaired.

GITHUBUSERNAME has been replaced with USERNAME in the description
of commit messages in docstrings to reflect the fact that
github usernames cannot be determined from the commit history
in the local git repository.