This adds a new command "docs" to cirrus which allows for building, packing as a tarball, and uploading Sphinx documentation to a remote server or uploading the tarball to a jenkins job to upload to and unpack the tarball on a remote server.
Functions for the docs commands are put in the documentation_utils.py module, and tests are added for it.

An uploader plugin 'file_server' is added for directly uploading the tarfile to a remote server, and a new function is added to jenkins for uploading a file with the jenkins build request.

Add credentials for accessing the file server to the default and keyring cred plugins, as well as updating the keyring creds test

Add requests-toolbelt as a requirement in order to use the MultipartEncoder supplied by the library. The MultipartEncoder supplies quick and easy functionality for encoding the tarfile so it can be uploaded to the Jenkins server. The choice to upload the file to Jenkins in favor of sending a job that will just do everything was made because it gives the user more flexibility and control in the preparation of their documentation.

Update the git cirrus build --docs function call to access build_docs() from the documentation_utils.py module. Add --docs flag to git cirrus release build and git cirrus release upload as shortcuts to the functionality of the git cirrus docs command. release build --docs will build the release tarball as well as compile the Sphinx docs and build the documentation tarball. release upload --docs will upload the release tarball and the doc tarball. To differentiate the release tarball upload method from the doc tarball upload method, the release continues to use the --plugin flag, and the docs specify the upload method in the cirrus.conf [doc] section.

add entry point for the git cirrus docs command